tests/wpt/web-platform-tests/custom-elements/registering-custom-elements/definition-construction-algorithm-no-interface-for-name.html

<!DOCTYPE html>
<html>
<head>
<title>If element interface for name doesn't exists then error must be thrown</title>
<meta name="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<meta name="assert" content="If NAME was provided and is not null and if element interface for the name and namespace does not exist or is an interface for a custom element, set ERROR to InvalidName and stop">
<link rel="help" href="http://www.w3.org/TR/custom-elements/#registering-custom-elements">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
</head>
<body>
<div id="log"></div>
<script>
test(function() {
    var doc = newHTMLDocument();
    doc.registerElement('x-a');
    assert_throws('NotSupportedError', function() {
        doc.registerElement('x-b', {extends: 'x-a'});
    }, 'Exception should be thrown in case of attempt to register ' +
        'a custom element which extends another custom element');
}, 'Exception should be thrown in case of attempt to register ' +
    'a custom element which extends another custom element');
</script>
</body>
</html>
